requirements.txt
chromadb
openai
langchain
uvicorn
fastapi
pypdf
langchain_community
python-dotenv
langchain-openai
langchain-chroma